+++ src/sll/include/sll/types.h
@@ -713,7 +713,7 @@ typedef union __SLL_OBJECT_DATA{
 
 typedef struct __SLL_OBJECT{
 	sll_ref_count_t rc;
-	sll_object_type_t t;
+	const sll_object_type_t t;
 	sll_object_data_t dt;
 } sll_object_t;